requirements.txt
# AI Stock Advisor - Requirements
# Python 3.10+ recommended

# Core AI Framework
crewai>=0.28.0
crewai-tools>=0.1.0
langchain-openai>=0.0.5

# LLM & Embeddings
openai>=1.12.0

# Data Sources
yfinance>=0.2.36

# Web UI
streamlit>=1.31.0
plotly>=5.18.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Optional: Better charts (matplotlib fallback)
matplotlib>=3.8.0
